requirements.txt
# Core FastAPI Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
starlette>=0.27.0

# Streamlit Frontend Dependencies
streamlit>=1.28.0

# HTTP Client for Frontend-Backend Communication
requests>=2.31.0

# Image Processing
Pillow>=10.0.0

# PDF Processing and OCR (Marker Library)
marker-pdf>=0.2.0

# CLI Interface
click>=8.1.0

# Logging and Error Handling
python-multipart>=0.0.6

# Database and Authentication
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.12.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Optional: For enhanced LLM processing
# google-generativeai>=0.3.0  # Uncomment if using Gemini API directly

# Development Dependencies (Optional)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0